Working Code
Added testers to each function, fixed overdivision of frequencies and inner product of torque amplitudes. Added displacement amplitude as a function of time to displacement amplitude as a function of frequency converter (dispAmpTF), as well as a displacement amplitude to torque amplitude converter (ampToPower). Still need to create a test for specFreqPower. Removed use of fakeDarkEPanalysis.