Rank Fixes and additional output
Fixed problems with not starting time at 1 with the fitter becoming degenerate, added outputs when the first fit fails and the design matrix without resonant terms is used. Also added a file to check fit values as a function of frequency (ampFreqCheck).

Need to figure out how to deal with the weights kicking in around 5e-3 Hz and higher, as well as the side frequencies being inflated.